Move Base GetColSums and GetAbsDxColSums to SimdBaseColSums.cpp
Extract Base::GetColSums and Base::GetAbsDxColSums from
SimdBaseStatistic.cpp into a dedicated SimdBaseColSums.cpp, and
register the new file in the VS2022 Base project.
