+++ prj/vs2022/Base.vcxproj
@@ -151,6 +151,7 @@
     <ClCompile Include="..\..\src\Simd\SimdBaseBgrToRgb.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseBgrToYuv.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseBinarization.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdBaseColSums.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseConditional.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseCopy.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseCpu.cpp" />
